chore(fe): rename useVocFilter hooks + split tag-master into ui/+api/
- useVocFilter.ts → useVocFilterContext.ts (context consumer)
- useVocFilters.ts → useVocFilterUrlState.ts (URL ↔ state sync)
- features/admin/tag-master/ flat → api/+ui/, ActionButton → TagMasterActionButton
- normalize tag contract import to @contracts/admin/tag alias

Per naming-conventions §5.2/§5.3. Adversarial review: PROCEED (Cand 2+4).
