+++ frontend/src/features/voc/list/index.ts
@@ -1,8 +1,8 @@
 export { VocFilterProvider, VocFilterContext } from './model/VocFilterContext';
 export type { VocFilters, VocFilterContextValue, VocStatus } from './model/VocFilterContext';
-export { useVocFilter } from './model/useVocFilter';
-export { useVocFilters } from './model/useVocFilters';
-export type { VocPageFilters } from './model/useVocFilters';
+export { useVocFilterContext } from './model/useVocFilterContext';
+export { useVocFilterUrlState } from './model/useVocFilterUrlState';
+export type { VocPageFilters } from './model/useVocFilterUrlState';
 export { VocListHeader } from './ui/VocListHeader';
 export { VOC_FIELD_LABELS } from '../constants';
 export { VOC_GRID_COLUMNS, VOC_GRID_PADDING_X } from './ui/vocGridLayout';
