+++ frontend/src/features/voc/list/model/useVocFilterContext.ts
@@ -2,8 +2,8 @@ import { useContext } from 'react';
 import { VocFilterContext } from './VocFilterContext';
 import type { VocFilterContextValue } from './VocFilterContext';
 
-export function useVocFilter(): VocFilterContextValue {
+export function useVocFilterContext(): VocFilterContextValue {
   const ctx = useContext(VocFilterContext);
-  if (!ctx) throw new Error('useVocFilter must be used within VOCFilterProvider');
+  if (!ctx) throw new Error('useVocFilterContext must be used within VOCFilterProvider');
   return ctx;
 }
